First cut
Has the ability to create a data structure describing a beaded
kumihimo braid, and to then create an SVG visualization of that braid
in an unbraided state.

The SVG generation will need to be moved from the data structure
classes to the visualization class, to separate the concerns. Doing so
will also allow us to control the SVG generation, and create
separate (e.g. braided) views.